requirements.txt
-e ./shoop
PyYAML==3.11
django-registration-redux==1.2
